Removes GTFS-aggregation funcionality
This responsibility has been moved from marduk to damu
