+++ helm/marduk/templates/configmap.yaml
@@ -137,8 +137,6 @@ data:
     otp.graph.build.remote.kubernetes.timeout=12000
     rutebanken.kubernetes.namespace={{ .Release.Namespace }}
     otp.graph.build.autoStartup={{ .Values.otp.graph.build.autoStartup }}
-
-    marduk.gtfs-aggregation-next.enabled={{ .Values.gtfsAggregationNext.enabled }}
 kind: ConfigMap
 metadata:
   name: marduk-application
